+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 * Add `setCountryCode` to `GooglePayRequest`
 * Add Google Pay support for Elo cards. 
+* Add `VenmoRequest` 
 * Breaking Changes
+  * Replace `VenmoClient#authorizeAccount()` methods with `VenmoClient#tokenizeVenmoAccount()`
   * Make `AmericanExpressRewardsBalance#fromJson()` package-private
   * Make `TYPE` and `API_RESOURCE_KEY` in `CardNonce` package-private
   * Make `CardNonce#fromJson()` methods package-private